Implement sqrt
The is a rather brutish port of the openlibm code. I'm not sure that all
the casts and wrapping operations reproduce the exact behavior of the C
code. The tests are failing with small differences to the libm results.

Parts of the code still need to be rewritten to be endian independent.